Refactor the code structure a little bit
