+++ sql/src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
@@ -1,1 +1,1 @@
-org.apache.spark.sql.sedona_sql.io.GeotiffImageFileFormat
+org.apache.spark.sql.sedona_sql.io.GeotiffFileFormat
