+++ sql/src/main/scala/org/apache/sedona/sql/UDF/Catalog.scala
@@ -21,10 +21,12 @@ package org.apache.sedona.sql.UDF
 import org.apache.spark.sql.catalyst.analysis.FunctionRegistry.FunctionBuilder
 import org.apache.spark.sql.expressions.{Aggregator, UserDefinedAggregateFunction}
 import org.apache.spark.sql.sedona_sql.expressions._
+import org.apache.spark.sql.sedona_sql.expressions.raster.{RS_AddBands, RS_Base64, RS_BitwiseAnd, RS_BitwiseOr, RS_Count, RS_DivideBands, RS_FetchRegion, RS_GetBand, RS_GreaterThan, RS_GreaterThanEqual, RS_HTML, RS_LessThan, RS_LessThanEqual, RS_LogicalDifference, RS_LogicalOver, RS_Mean, RS_Mode, RS_Modulo, RS_MultiplyBands, RS_MultiplyFactor, RS_NormalizedDifference, RS_SquareRoot, RS_SubtractBands}
 import org.locationtech.jts.geom.Geometry
 
 object Catalog {
   val expressions: Seq[FunctionBuilder] = Seq(
+    // Expression for vectors
     ST_PointFromText,
     ST_PolygonFromText,
     ST_LineStringFromText,
@@ -84,6 +86,8 @@ object Catalog {
     ST_LineInterpolatePoint,
     ST_SubDivideExplode,
     ST_SubDivide,
+
+    // Expression for rasters
     RS_NormalizedDifference,
     RS_Mean,
     RS_Mode,
@@ -105,7 +109,8 @@ object Catalog {
     RS_SquareRoot,
     RS_LogicalDifference,
     RS_LogicalOver,
-    RS_Base64
+    RS_Base64,
+    RS_HTML
   )
 
   val aggregateExpressions: Seq[Aggregator[Geometry, Geometry, Geometry]] = Seq(
